SKALE-4769 Update all tests to create new nodes and schains
