+++ pytest.ini
@@ -1,5 +1,5 @@
 [pytest]
-log_cli = 1
+log_cli = 0
 log_cli_level = INFO
 log_cli_format = %(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)
 log_cli_date_format=%Y-%m-%d %H:%M:%S
