+++ tests/schain_config/generator_test.py
@@ -1,10 +1,12 @@
 import skale.contracts.manager.nodes as nodes
 from skale.schain_config.generator import get_nodes_for_schain, get_schain_nodes_with_schains
-from tests.constants import (DEFAULT_SCHAIN_NAME, MIN_NODES_IN_SCHAIN)
+from skale.utils.contracts_provision.main import create_clean_schain
+from tests.constants import MIN_NODES_IN_SCHAIN
 
 
 def test_get_nodes_for_schain(skale):
-    schain_nodes = get_nodes_for_schain(skale, DEFAULT_SCHAIN_NAME)
+    schain_name = create_clean_schain(skale)
+    schain_nodes = get_nodes_for_schain(skale, schain_name)
     fields_with_id = nodes.FIELDS.copy()
     fields_with_id.append('id')
 
@@ -13,6 +15,7 @@ def test_get_nodes_for_schain(skale):
 
 
 def test_get_schain_nodes_with_schains(skale):
-    nodes_with_schains = get_schain_nodes_with_schains(skale, DEFAULT_SCHAIN_NAME)
+    schain_name = create_clean_schain(skale)
+    nodes_with_schains = get_schain_nodes_with_schains(skale, schain_name)
     assert isinstance(nodes_with_schains[0]['schains'], list)
     assert isinstance(nodes_with_schains[0]['schains'][0]['mainnetOwner'], str)
